src/backend/routers/kinect_dashboard.py
#!/usr/bin/env python3
"""
VokaFlow - Router de Dashboard Kinect
Integra funcionalidades de Kinect con el dashboard web
"""

import os
import logging
import asyncio
import json
from datetime import datetime
from typing import Optional, Dict, Any, List
from fastapi import APIRouter, HTTPException, WebSocket, WebSocketDisconnect
from fastapi.responses import StreamingResponse
from pydantic import BaseModel
import cv2
import numpy as np

# Configuración de logging
logger = logging.getLogger("vokaflow.kinect_dashboard")

# Router
router = APIRouter()

# Modelos Pydantic
class KinectStatus(BaseModel):
    connected: bool
    camera: bool
    microphone: bool
    audio: bool
    depth_sensor: bool
    skeleton_tracking: bool
    face_detection: bool
    gesture_recognition: bool

class KinectMetrics(BaseModel):
    fps: float
    depth_range: str
    detected_users: int
    audio_level: float
    gesture_count: int
    face_count: int

# Conexiones WebSocket activas
active_connections: List[WebSocket] = []

@router.get("/status", response_model=KinectStatus)
async def get_kinect_status():
    """Obtiene el estado actual del sistema Kinect"""
    try:
        # Simulación de estado de Kinect
        return KinectStatus(
            connected=True,
            camera=True,
            microphone=True,
            audio=True,
            depth_sensor=True,
            skeleton_tracking=True,
            face_detection=True,
            gesture_recognition=True
        )
    except Exception as e:
        logger.error(f"Error obteniendo estado de Kinect: {e}")
        raise HTTPException(status_code=500, detail="Error obteniendo estado de Kinect")

@router.get("/metrics", response_model=KinectMetrics)
async def get_kinect_metrics():
    """Obtiene métricas en tiempo real del sistema Kinect"""
    try:
        # Simulación de métricas
        return KinectMetrics(
            fps=30.0,
            depth_range="0.8m - 4.0m",
            detected_users=2,
            audio_level=75.0,
            gesture_count=5,
            face_count=2
        )
    except Exception as e:
        logger.error(f"Error obteniendo métricas de Kinect: {e}")
        raise HTTPException(status_code=500, detail="Error obteniendo métricas de Kinect")

@router.get("/video/stream")
async def get_video_stream():
    """Stream de video en tiempo real desde Kinect"""
    try:
        def generate_frames():
            # Generar frames simulados
            for i in range(1000):
                # Frame simulado
                frame = np.random.randint(0, 255, (480, 640, 3), dtype=np.uint8)
                _, buffer = cv2.imencode('.jpg', frame)
                frame_bytes = buffer.tobytes()
                
                yield (b'--frame\r\n'
                       b'Content-Type: image/jpeg\r\n\r\n' + frame_bytes + b'\r\n')
        
        return StreamingResponse(
            generate_frames(),
            media_type="multipart/x-mixed-replace; boundary=frame"
        )
    except Exception as e:
        logger.error(f"Error iniciando stream de video: {e}")
        raise HTTPException(status_code=500, detail="Error iniciando stream de video")

@router.post("/video/toggle")
async def toggle_video(config: Dict[str, bool]):
    """Activa/desactiva el video de Kinect"""
    try:
        enabled = config.get('enabled', False)
        return {"success": True, "video_enabled": enabled}
    except Exception as e:
        logger.error(f"Error toggling video: {e}")
        raise HTTPException(status_code=500, detail="Error toggling video")

@router.post("/audio/toggle")
async def toggle_audio(config: Dict[str, bool]):
    """Activa/desactiva el audio de Kinect"""
    try:
        enabled = config.get('enabled', False)
        return {"success": True, "audio_enabled": enabled}
    except Exception as e:
        logger.error(f"Error toggling audio: {e}")
        raise HTTPException(status_code=500, detail="Error toggling audio")

@router.post("/recording/start")
async def start_recording():
    """Inicia grabación de video y audio"""
    try:
        filename = f"kinect_recording_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
        return {"success": True, "recording": True, "filename": filename}
    except Exception as e:
        logger.error(f"Error iniciando grabación: {e}")
        raise HTTPException(status_code=500, detail="Error iniciando grabación")

@router.post("/recording/stop")
async def stop_recording():
    """Detiene grabación de video y audio"""
    try:
        return {"success": True, "recording": False, "saved_file": "recording_saved.mp4"}
    except Exception as e:
        logger.error(f"Error deteniendo grabación: {e}")
        raise HTTPException(status_code=500, detail="Error deteniendo grabación")

@router.post("/calibrate")
async def calibrate_kinect():
    """Calibra el sistema Kinect"""
    try:
        return {"success": True, "calibration_result": "Calibración completada"}
    except Exception as e:
        logger.error(f"Error calibrando Kinect: {e}")
        raise HTTPException(status_code=500, detail="Error calibrando Kinect")

@router.websocket("/ws")
async def websocket_endpoint(websocket: WebSocket):
    """WebSocket para actualizaciones en tiempo real"""
    await websocket.accept()
    active_connections.append(websocket)
    
    try:
        while True:
            data = {
                "type": "update",
                "metrics": {
                    "fps": 30.0,
                    "detected_users": 2,
                    "audio_level": 75.0,
                    "gesture_count": 5,
                    "face_count": 2
                },
                "status": {
                    "connected": True,
                    "camera": True,
                    "microphone": True
                },
                "timestamp": datetime.now().isoformat()
            }
            
            await websocket.send_text(json.dumps(data))
            await asyncio.sleep(1)
            
    except WebSocketDisconnect:
        active_connections.remove(websocket)
    except Exception as e:
        logger.error(f"Error en WebSocket: {e}")
        if websocket in active_connections:
            active_connections.remove(websocket)

@router.get("/service/status")
async def kinect_dashboard_service_status():
    """Verifica el estado del servicio de dashboard Kinect"""
    return {
        "status": "operational",
        "version": "1.0.0",
        "kinect_available": True,
        "kinect_connected": True,
        "active_connections": len(active_connections),
        "components": {
            "video_stream": "operational",
            "depth_stream": "operational",
            "audio_system": "operational",
            "websocket": "operational"
        }
    }
